+++ bpstudio-20210931/core/Models/AppFailBackRequest.cs
@@ -1,0 +1,21 @@
+/**
+ *
+ */
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.BPStudio20210931.Models
+{
+    public class AppFailBackRequest : TeaModel {
+        [NameInMap("ApplicationId")]
+        [Validation(Required=false)]
+        public string ApplicationId { get; set; }
+
+    }
+
+}