+++ bpstudio-20210931/core/Models/CreateApplicationResponse.cs
@@ -10,15 +10,15 @@ namespace AlibabaCloud.SDK.BPStudio20210931.Models
 {
     public class CreateApplicationResponse : TeaModel {
         [NameInMap("headers")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public Dictionary<string, string> Headers { get; set; }
 
         [NameInMap("statusCode")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public int? StatusCode { get; set; }
 
         [NameInMap("body")]
-        [Validation(Required=true)]
+        [Validation(Required=false)]
         public CreateApplicationResponseBody Body { get; set; }
 
     }