+++ bpstudio-20210931/core/Models/GetPotentialFailZonesRequest.cs
@@ -1,0 +1,22 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.BPStudio20210931.Models
+{
+    public class GetPotentialFailZonesRequest : TeaModel {
+        [NameInMap("IsPlanId")]
+        [Validation(Required=false)]
+        public bool? IsPlanId { get; set; }
+
+        [NameInMap("ObjectId")]
+        [Validation(Required=false)]
+        public string ObjectId { get; set; }
+
+    }
+
+}