+++ bpstudio-20210931/ChangeLog.md
@@ -1,3 +1,14 @@
+2024-01-30 Version: 2.0.0
+- Support API AppFailBack.
+- Support API AppFailOver.
+- Support API GetFoTaskStatus.
+- Support API GetPotentialFailZones.
+- Support API InitAppFailOver.
+- Update API ListApplicationadd ResourceId param.
+add TemplateId param.
+- Update API ListFoCreatedAppsupdate response param.
+
+
 2024-01-10 Version: 1.1.1
 - Generated csharp 2021-09-31 for BPStudio.